pyproject.toml
[tool.poetry]
name = "drf_crud"
version = "0.1.0"
description = ""
authors = ["tsorokin <tsorokin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0"
djangorestframework = "^3.13.1"
apiview = "^1.3.25"
psycopg2-binary = "^2.9.2"
psycopg2 = "^2.9.2"
dj-config-url = "^0.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
